Update Nashorn
git-svn-id: http://svn.code.sf.net/p/htmlunit/code/trunk@12469 5f5364db-9458-4db8-a492-e30667be6df6

+++ js/src/main/java/com/gargoylesoftware/js/nashorn/internal/codegen/ConstantData.java
@@ -174,7 +174,7 @@ final class ConstantData {
         final Integer value = stringMap.get(string);
 
         if (value != null) {
-            return value;
+            return value.intValue();
         }
 
         constants.add(string);
@@ -203,7 +203,7 @@ final class ConstantData {
         final Integer value = objectMap.get(entry);
 
         if (value != null) {
-            return value;
+            return value.intValue();
         }
 
         constants.add(object);
